requirements.txt
pandas
psycopg2-binary
mysql-connector-python
sqlalchemy
pymongo
matplotlib
seaborn
streamlit
streamlit-autorefresh
faker
pymysql
cryptography